final/requirements.txt
numpy==1.16.3
Pillow==6.0.0
polib==1.1.0
scipy==1.2.0
six==1.12.0
torch==1.1.0
torchvision==0.3.0
tqdm==4.32.1
